Fix all compiler warnings across dirmaster, inet_client, smtp, http_client
- dirmaster.h: Increase date string buffers from 20 to 64 bytes (format-truncation)
- dirmaster.c: Cast size_decimals to size_t for sizeof comparison (sign-compare)
- inet_client.c, smtp.c, http_client.c: Cast tag to unsigned int for AFC_TAG_END comparison
- http_client.c: Cast afc_string_max() to int for comparisons with int variables
- http_client.c: Handle fgets() return value to fix unused-result warning

+++ src/dirmaster.c
@@ -1081,7 +1081,7 @@ static void afc_dirmaster_internal_size2string(DirMaster *dm, char *dest, unsign
 		if (dm->size_decimals)
 		{
 			snprintf(buf, 20, "%lu", size - (rbase * rsize));
-			if (dm->size_decimals < sizeof(buf)) buf[dm->size_decimals] = 0;
+			if ((size_t)dm->size_decimals < sizeof(buf)) buf[dm->size_decimals] = 0;
 			snprintf(dest, 20, "%d.%.*s %s", rsize, dm->size_decimals, buf, afc_dirmaster_size_bases[count]);
 		}
 		else
